Move all unit tests to testing directory
