+++ packages/lesswrong/testing/client.tests.js
@@ -1,3 +1,3 @@
-import '../components/comments/CommentsItem/_comments-unit-tests.js';
-import '../components/posts/timeframeUtils.tests.js';
+import './commentsItem.tests.js';
+import './timeframeUtils.tests.js';
 import './moderation/moderation.frontend.tests.js';
