+++ packages/lesswrong/testing/notifications.tests.js
@@ -1,10 +1,10 @@
 import { chai } from 'meteor/practicalmeteor:chai';
 import chaiAsPromised from 'chai-as-promised';
-import { createDummyUser, createDummyPost, createDummyConversation, createDummyMessage } from '../../../testing/utils.js'
-import { performSubscriptionAction } from '../subscriptions/mutations.js';
+import { createDummyUser, createDummyPost, createDummyConversation, createDummyMessage } from './utils.js'
+import { performSubscriptionAction } from '../lib/collections/subscriptions/mutations.js';
 
 import Users from 'meteor/vulcan:users';
-import Notifications from './collection.js';
+import Notifications from '../lib/collections/notifications/collection.js';
 //import Comments from '../comments/collection.js';
 import { waitUntilCallbacksFinished } from 'meteor/vulcan:core';
