+++ packages/lesswrong/testing/timeframeUtils.tests.js
@@ -1,4 +1,4 @@
-import { getDateRange } from './timeframeUtils'
+import { getDateRange } from '../components/posts/timeframeUtils'
 
 describe('getDateRange', () => {
   it('handles days', () => {
